+++ config/inertia.php
@@ -139,4 +139,6 @@ return [
 
     ],
 
+    'use_script_element_for_initial_page' => true,
+
 ];